minor website update

+++ src/slashCommands/moderation/role.js
@@ -17,7 +17,10 @@ module.exports = {
     .addBooleanOption((option) => option.setName("remove").setDescription("Remove role or not")))
   .addSubCommand((subcommand) => subcommand.setName("add").setDescription("Adds a role to a user.")
     .addUserOption((option) => option.setName("user").setDescription("The user to add the role to.").setRequired(true))
-    .addRoleOption((option) => option.setName("role").setDescription("The role to add."))),
+    .addRoleOption((option) => option.setName("role").setDescription("The role to add.").setRequired(true)))
+  .addSubCommand((subcommand) => subcommand.setName("remove").setDescription("Removes a role from a user.")
+    .addUserOption((option) => option.setName("user").setDescription("The user to remove the role from.").setRequired(true))
+    .addRoleOption((option) => option.setName("role").setDescription("The role to remove.").setRequired(true))),
   async execute(interaction) {
     if(!interaction.member.permissions.has("MANAGE_ROLES")) {
       return interaction.reply({ content: `You do not have permission to use this command.`, ephemeral: true });
@@ -45,27 +48,26 @@ module.exports = {
         })
         .setDescription(`${fail} | Please provide a valid role!`)
         .setTimestamp()
-        .setFooter({ text: "https://mee8.ml" })
+        .setFooter({ text: "https://6c7f021b-2514-4460-9d5a-64060cec1990-00-30w9y136gg7mt.riker.replit.dev" })
         .setColor(interaction.client.color.red)
         return interaction.reply({ embeds: [rolenotfound], ephmeral: true })
-        .then(async () => {
-          if (logging && logging.moderation.delete_reply === "true") {
-            setTimeout(() => {
-              interaction.deleteReply().catch(() => {});
-            }, 5000)
-          }
-        })
-        .catch(() => {});
       } else {
         if(removerole === false) {
-          interaction.reply({ content: `${success} | Adding ${role} to ${fetchedMembers.size} members. This may take a while!` }).then(() => {
+          interaction.reply({
+            embeds: [
+              new MessageEmbed()
+              .setDescription(`${success} | Adding ${role} to ${interaction.guild.members.cache.size} members. This may take a while!`)
+              .setColor(interaction.client.color.green)
+            ]
+          }).then(() => {
             interaction.guild.members.cache.forEach((member) => member.roles.add(role, [
-              `Role Add / Responsible User: ${interaction.user.tag}`,
-            ]));
+              `Role Add / Responsible User: ${interaction.user.tag}`
+            ]))
+          }).then(() => {
             const embed = new MessageEmbed()
             .setDescription(`${success} | Added **${role}** to **${interaction.guild.members.cache.size.size}** members.`)
             .setColor(interaction.client.color.green)
-            interaction.channel.send({ embeds: [embed] })
+            interaction.editReply({ embeds: [embed] })
             .then(async () => {
               if (logging && logging.moderation.delete_reply === "true") {
                 setTimeout(() => {
@@ -76,14 +78,21 @@ module.exports = {
             .catch(() => {})
           });
         } else {
-          interaction.reply({ content: `${success} | Removing ${role} from ${interaction.guild.members.cache.size} members. This may take a while!` }).then(() => {
+          interaction.reply({
+            embeds: [
+              new MessageEmbed()
+              .setDescription(`${success} | Removing ${role} from ${interaction.guild.members.cache.size} members. This may take a while!`)
+              .setColor(interaction.client.color.green)
+            ]
+          }).then(() => {
             interaction.guild.members.cache.forEach((member) => member.roles.remove(role, [
               `Role Remove / Responsible User: ${interaction.user.tag}`,
-            ]));
+            ]))
+          }).then(() => {
             const embed = new MessageEmbed()
-            .setDescription(`${success} | Removed **${role}** from **${fetchedMembers.size}** members.`)
+            .setDescription(`${success} | Removed **${role}** from **${interaction.guild.members.cache.size}** members.`)
             .setColor(interaction.client.color.green)
-            interaction.channel.send({ embeds: [embed] })
+            interaction.editReply({ embeds: [embed] })
             .then(async () => {
               if (logging && logging.moderation.delete_reply === "true") {
                 setTimeout(() => {
@@ -115,27 +124,26 @@ module.exports = {
         })
         .setDescription(`${fail} | Please provide a valid role!`)
         .setTimestamp()
-        .setFooter({ text: "https://mee8.ml" })
+        .setFooter({ text: "https://6c7f021b-2514-4460-9d5a-64060cec1990-00-30w9y136gg7mt.riker.replit.dev" })
         .setColor(interaction.client.color.red)
         return interaction.reply({ embeds: [rolenotfound], ephmeral: true })
-        .then(async () => {
-          if (logging && logging.moderation.delete_reply === "true") {
-            setTimeout(() => {
-              interaction.deleteReply().catch(() => {});
-            }, 5000)
-          }
-        })
-        .catch(() => {});
       } else {
         if(removerole === false) {
-          interaction.reply({ content: `${success} | Adding ${role} to ${interaction.guild.members.cache.filter((m) => m.user.bot).size} bots. This may take a while!` }).then(() => {
+          interaction.reply({
+            embeds: [
+              new MessageEmbed()
+              .setDescription(`${success} | Adding ${role} to ${interaction.guild.members.cache.filter((m) => m.user.bot).size} bots. This may take a while!`)
+              .setColor(interaction.client.color.green)
+            ]
+          }).then(() => {
             interaction.guild.members.cache.filter((m) => m.user.bot).forEach((member) => member.roles.add(role, [
               `Role Add / Responsible User: ${interaction.user.tag}`,
-            ]));
+            ]))
+          }).then(() => {
             const embed = new MessageEmbed()
             .setDescription(`${success} | Added **${role}** to **${interaction.guild.members.cache.filter((m) => m.user.bot).size}** bots.`)
             .setColor(interaction.client.color.green)
-            interaction.channel.send({ embeds: [embed] })
+            interaction.editReply({ embeds: [embed] })
             .then(async () => {
               if (logging && logging.moderation.delete_reply === "true") {
                 setTimeout(() => {
@@ -146,14 +154,21 @@ module.exports = {
             .catch(() => {})
           });
         } else {
-          interaction.reply({ content: `${success} | Removing ${role} from ${interaction.guild.members.cache.filter((m) => m.user.bot).size} bots. This may take a while!` }).then(() => {
+          interaction.reply({
+            embeds: [
+              new MessageEmbed()
+              .setDescription(`${success} | Removing ${role} from ${interaction.guild.members.cache.filter((m) => m.user.bot).size} bots. This may take a while!`)
+              .setColor(interaction.client.color.green)
+            ]
+          }).then(() => {
             interaction.guild.members.cache.filter((m) => m.user.bot).forEach((member) => member.roles.remove(role, [
               `Role Remove / Responsible User: ${interaction.user.tag}`,
-            ]));
+            ]))
+          }).then(() => {
             const embed = new MessageEmbed()
             .setDescription(`${success} | Removed **${role}** from **${interaction.guild.members.cache.filter((m) => m.user.bot).size}** bots.`)
             .setColor(interaction.client.color.green)
-            interaction.channel.send({ embeds: [embed] })
+            interaction.editReply({ embeds: [embed] })
             .then(async () => {
               if (logging && logging.moderation.delete_reply === "true") {
                 setTimeout(() => {
@@ -185,27 +200,26 @@ module.exports = {
         })
         .setDescription(`${fail} | Please provide a valid role!`)
         .setTimestamp()
-        .setFooter({ text: "https://mee8.ml" })
+        .setFooter({ text: "https://6c7f021b-2514-4460-9d5a-64060cec1990-00-30w9y136gg7mt.riker.replit.dev" })
         .setColor(interaction.client.color.red)
         return interaction.reply({ embeds: [rolenotfound], ephmeral: true })
-        .then(async () => {
-          if (logging && logging.moderation.delete_reply === "true") {
-            setTimeout(() => {
-              interaction.deleteReply().catch(() => {});
-            }, 5000)
-          }
-        })
-        .catch(() => {});
       } else {
         if(removerole === false) {
-          interaction.reply({ content: `${success} | Adding ${role.name} to ${interaction.guild.members.cache.filter((m) => !m.user.bot).size} humans. This may take a while!` }).then(() => {
+          interaction.reply({
+            embeds: [
+              new MessageEmbed()
+              .setDescription(`${success} | Adding ${role.name} to ${interaction.guild.members.cache.filter((m) => !m.user.bot).size} humans. This may take a while!`)
+              .setColor(interaction.client.color.green)
+            ]
+          }).then(() => {
             interaction.guild.members.cache.filter((m) => !m.user.bot).forEach((member) => member.roles.add(role, [
               `Role Add / Responsible User: ${interaction.user.tag}`,
-            ]));
+            ]))
+          }).then(() => {
             const embed = new MessageEmbed()
             .setDescription(`${success} | Added **${role}** to **${interaction.guild.members.cache.filter((m) => !m.user.bot).size}** humans.`)
             .setColor(interaction.client.color.green)
-            interaction.channel.send({ embeds: [embed] })
+            interaction.editReply({ embeds: [embed] })
             .then(async () => {
               if (logging && logging.moderation.delete_reply === "true") {
                 setTimeout(() => {
@@ -216,14 +230,21 @@ module.exports = {
             .catch(() => {})
           });
         } else {
-          interaction.reply({ content: `${success} | Removing ${role.name} from ${interaction.guild.members.cache.filter((m) => !m.user.bot).size} humans. This may take a while!` }).then(() => {
+          interaction.reply({
+            embeds: [
+              new MessageEmbed()
+              .setDescription(`${success} | Removing ${role.name} from ${interaction.guild.members.cache.filter((m) => !m.user.bot).size} humans. This may take a while!`)
+              .setColor(interaction.client.color.green)
+            ]
+          }).then(() => {
             interaction.guild.members.cache.filter((m) => !m.user.bot).forEach((member) => member.roles.remove(role, [
               `Role Remove / Responsible User: ${interaction.user.tag}`,
             ]))
+          }).then(() => {
             const embed = new MessageEmbed()
             .setDescription(`${success} | Removed **${role}** from **${interaction.guild.members.cache.filter((m) => !m.user.bot).size}** humans.`)
             .setColor(interaction.client.color.green)
-            interaction.channel.send({ embeds: [embed] })
+            interaction.editReply({ embeds: [embed] })
             .then(async () => {
               if (logging && logging.moderation.delete_reply === "true") {
                 setTimeout(() => {
@@ -254,17 +275,9 @@ module.exports = {
         })
         .setDescription(`${fail} | Please provide a valid role!`)
         .setTimestamp()
-        .setFooter({ text: "https://mee8.ml" })
+        .setFooter({ text: "https://6c7f021b-2514-4460-9d5a-64060cec1990-00-30w9y136gg7mt.riker.replit.dev" })
         .setColor(interaction.client.color.red)
         return interaction.reply({ embeds: [rolenotfound], ephmeral: true })
-        .then(async () => {
-          if (logging && logging.moderation.delete_reply === "true") {
-            setTimeout(() => {
-              interaction.deleteReply().catch(() => {});
-            }, 5000)
-          }
-        })
-        .catch(() => {});
       } else {
         if(member.roles.cache.has(role.id)) {
           let alreadyhasrole = new MessageEmbed()
@@ -274,16 +287,9 @@ module.exports = {
           })
           .setDescription(`${fail} | ${member} already has the role ${role}.`)
           .setTimestamp()
-          .setFooter({ text: "https://mee8.ml" })
+          .setFooter({ text: "https://6c7f021b-2514-4460-9d5a-64060cec1990-00-30w9y136gg7mt.riker.replit.dev" })
           .setColor(interaction.client.color.red)
           return interaction.reply({ embeds: [alreadyhasrole], ephmeral: true })
-          .then(async () => {
-            if (logging && logging.moderation.delete_reply === "true") {
-              setTimeout(() => {
-                interaction.deleteReply().catch(() => {});
-              }, 5000)
-            }
-          })
         } else {
           member.roles.add(role, [
             `Role Add / Responsible User: ${interaction.user.tag}`,
@@ -308,16 +314,73 @@ module.exports = {
             })
             .setDescription(`${fail} | The role is possibly higher than me or you. Please move my role above the role and try again!`)
             .setTimestamp()
-            .setFooter({ text: "https://mee8.ml" })
+            .setFooter({ text: "https://6c7f021b-2514-4460-9d5a-64060cec1990-00-30w9y136gg7mt.riker.replit.dev" })
             .setColor(interaction.client.color.red)
             return interaction.reply({ embeds: [botrolepossiblylow], ephmeral: true })
-            .then(async () => {
+          })
+        }
+      }
+    } else if (interaction.options.getSubcommand() === "remove") {
+      let member = interaction.options.getMember("user");
+      let role = interaction.options.getRole("role") || interaction.guild.roles.cache.get(role) || interaction.guild.roles.cache.find((rl) => rl.name.toLowerCase() === role.slice(1).join(" ").toLowerCase());
+      let reason = `The current feature doesn't need a reason.`;
+      if(!reason) {
+        reason = `No reason provided.`;
+      }
+      if(reason.length > 1024) {
+        reason = reason.slice(0, 1021) + "...";
+      }
+
+      if(!role) {
+        const rolenotfound = new MessageEmbed()
+        .setAuthor({
+          name: `${interaction.user.tag}`,
+          iconURL: interaction.user.displayAvatarURL({ dynamic: true })
+        })
+        .setDescription(`${fail} | Please provide a valid role!`)
+          .setTimestamp()
+          .setFooter({ text: "https://6c7f021b-2514-4460-9d5a-64060cec1990-00-30w9y136gg7mt.riker.replit.dev" })
+        .setColor(interaction.client.color.red)
+        return interaction.reply({ embeds: [rolenotfound], ephemeral: true })
+      } else {
+        if(!member.roles.cache.has(role.id)) {
+          const nothasrole = new MessageEmbed()
+          .setAuthor({
+            name: `${interaction.user.tag}`,
+            iconURL: interaction.user.displayAvatarURL({ dynamic: true })
+          })
+          .setDescription(`${fail} | ${member} doesn't have the role ${role}!`)
+          .setTimestamp()
+          .setFooter({ text: "https://6c7f021b-2514-4460-9d5a-64060cec1990-00-30w9y136gg7mt.riker.replit.dev" })
+          .setColor(interaction.client.color.red)
+          return interaction.reply({ embeds: [nothasrole], ephemeral: true })
+        } else {
+          member.roles.remove(role, [
+            `Role Remove / Responsible User: ${interaction.user.tag}`
+          ]).then(() => {
+            const embed = new MessageEmbed()
+            .setDescription(`${success} | Removed ${role} from ${member}.`)
+            .setColor(interaction.client.color.green)
+            interaction.reply({ embeds: [embed] })
+            .then(() => {
               if (logging && logging.moderation.delete_reply === "true") {
                 setTimeout(() => {
                   interaction.deleteReply().catch(() => {});
                 }, 5000)
               }
             })
+            .catch(() => {});
+          }).catch(() => {
+            let botrolepossiblylow = new MessageEmbed()
+            .setAuthor({
+              name: `${interaction.user.tag}`,
+              iconURL: interaction.user.displayAvatarURL({ dynamic: true })
+            })
+            .setDescription(`${fail} | The role is possibly higher than me or you. Please move my role above the role and try again.`)
+            .setTimestamp()
+            .setFooter({ text: `https://6c7f021b-2514-4460-9d5a-64060cec1990-00-30w9y136gg7mt.riker.replit.dev` })
+            .setColor(interaction.client.color.red)
+            return interaction.reply({ embeds: [botrolepossiblylow], ephemeral: true });
           })
         }
       }
